+++ sdk/storage/azure_storage_blob/tests/blob_client.rs
@@ -238,18 +238,16 @@ async fn test_set_blob_metadata(ctx: TestContext) -> Result<(), Box<dyn Error>> 
 
     // Set Metadata With Values
     let update_metadata = HashMap::from([("updated".to_string(), "values".to_string())]);
-    let set_metadata_options = BlobClientSetMetadataOptions {
-        metadata: Some(update_metadata.clone()),
-        ..Default::default()
-    };
-    blob_client.set_metadata(Some(set_metadata_options)).await?;
+    blob_client
+        .set_metadata(update_metadata.clone(), None)
+        .await?;
     // Assert
     let response = blob_client.get_properties(None).await?;
     let response_metadata = response.metadata()?;
     assert_eq!(update_metadata, response_metadata);
 
     // Set Metadata No Values (Clear Metadata)
-    blob_client.set_metadata(None).await?;
+    blob_client.set_metadata(HashMap::new(), None).await?;
     // Assert
     let response = blob_client.get_properties(None).await?;
     let response_metadata = response.metadata()?;
@@ -361,12 +359,9 @@ async fn test_leased_blob_operations(ctx: TestContext) -> Result<(), Box<dyn Err
         .await?;
 
     let update_metadata = HashMap::from([("updated".to_string(), "values".to_string())]);
-    let set_metadata_options = BlobClientSetMetadataOptions {
-        metadata: Some(update_metadata.clone()),
-        lease_id: Some(lease_id.clone()),
-        ..Default::default()
-    };
-    blob_client.set_metadata(Some(set_metadata_options)).await?;
+    blob_client
+        .set_metadata(update_metadata.clone(), None)
+        .await?;
 
     let set_tier_options = BlobClientSetTierOptions {
         lease_id: Some(lease_id.clone()),